Add files via upload
This is Artificial Intelligence Assignment.
The assignment is about introducing Uniform Cost Search Algorithm to Schedule the process and calculate the Average waiting time for the process. Here we are comparing the process scheduling Round Robin with UCS and comparing thee results.We found that the average waiting time given UCS is lesser and better than Round Robin. Similarly, the average Turn Around Time is efficient than Round Robin. The Execution time is also comparatively lesser for Uniform Cost Search than Round Robin. So, we can claim that we can use Uniform cost Search as an alternative for Round Robin. The input file used for the project is alo attached which consists of 3 columns with Process id, Arrival Time, Burst Time for 60 processes. Thank you!